movies_database_api/movies_database_api_requirements.txt
cursor==1.3.5
halo==0.0.29
log-symbols==0.0.14
parquet-tools==0.2.13
pyarrow==11.0.0
spinners==0.0.24
thrift==0.13.0
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.7
pydantic==2.4.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
annotated-types==0.5.0
anyio==3.7.1
boto3==1.34.31
botocore==1.34.31
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
cursor==1.3.5
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
halo==0.0.29
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
log-symbols==0.0.14
MarkupSafe==2.1.3
numpy==1.24.4
orjson==3.9.7
pandas==1.5.3
parquet-tools==0.2.13
psycopg2-binary==2.9.9
pyarrow==11.0.0
pycparser==2.21
pydantic==2.4.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.4
PyYAML==6.0.1
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
spinners==0.0.24
SQLAlchemy==2.0.25
starlette==0.27.0
tabulate==0.8.10
termcolor==2.4.0
thrift==0.13.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
